Adds handler interface and get and head implementations. Code is duplicated.
