+++ src/test/java/RouterTest.java
@@ -1,0 +1,21 @@
+import HTTPServer.Handler;
+import junit.framework.TestCase;
+import org.hamcrest.core.IsInstanceOf;
+
+import java.util.HashMap;
+
+/**
+ * Created by jphoenix on 8/23/16.
+ */
+public class RouterTest extends TestCase {
+    public void test_it_routes_to_get_handler() throws Exception {
+        assertEquals(1, 1);
+//        HashMap params = new HashMap();
+//        params.put("action", "GET");
+//        Router router = new Router();
+//
+//        router.route(params);
+//
+//        IsInstanceOf(getHandler, GetHandler);
+    }
+}
